feat : API 수정 및 postgreSQL로 변경

+++ src/main/kotlin/com/beyoureyes/beyoureyes/entity/DailyFood.kt
@@ -2,17 +2,17 @@ package com.beyoureyes.beyoureyes.entity
 
 import java.time.LocalDateTime
 
-data class DailyFood (
+data class DailyFood(
     val logId: Long? = null,
     val userId: Long,
     val dateTime: LocalDateTime = LocalDateTime.now(),
     val foodPhoto: String,
-    val calories: Double = 0.0,
-    val carbohydrates: Double = 0.0,
-    val protein: Double = 0.0,
-    val fat: Double = 0.0,
-    val cholesterol: Double = 0.0,
-    val sodium: Double = 0.0,
-    val sugar: Double = 0.0,
-    val saturatedFat: Double = 0.0
+    val calories: Int = 0,
+    val carbohydrates: Int = 0,
+    val protein: Int = 0,
+    val fat: Int = 0,
+    val cholesterol: Int = 0,
+    val sodium: Int = 0,
+    val sugar: Int = 0,
+    val saturatedFat: Int = 0
 )
